Add `username` validation business logic in controller and remove `user::create::username` component
